+++ napari/plugins/_tests/test_save_layers.py
@@ -1,13 +1,8 @@
 import os
 from napari.plugins.io import save_layers
-from napari.plugins._tests.fixtures.layer_data import (  # noqa: F401
-    layer_data_and_types,
-)
 
 
-def test_save_layer_single_named_plugin(
-    tmpdir, layer_data_and_types  # noqa: F811
-):
+def test_save_layer_single_named_plugin(tmpdir, layer_data_and_types):
     """Test saving a single layer with a named plugin."""
     layers, _, _, filenames = layer_data_and_types
 
@@ -24,9 +19,7 @@ def test_save_layer_single_named_plugin(
         assert os.path.isfile(path)
 
 
-def test_save_layer_single_no_named_plugin(
-    tmpdir, layer_data_and_types  # noqa: F811
-):
+def test_save_layer_single_no_named_plugin(tmpdir, layer_data_and_types):
     """Test saving a single layer without naming plugin."""
     # make writer builtin plugins get called first
     from napari.plugins import plugin_manager
@@ -50,9 +43,7 @@ def test_save_layer_single_no_named_plugin(
         assert os.path.isfile(path)
 
 
-def test_save_layer_multiple_named_plugin(
-    tmpdir, layer_data_and_types  # noqa: F811
-):
+def test_save_layer_multiple_named_plugin(tmpdir, layer_data_and_types):
     """Test saving multiple layers with a named plugin."""
     layers, _, _, filenames = layer_data_and_types
 
@@ -76,9 +67,7 @@ def test_save_layer_multiple_named_plugin(
     assert set(os.listdir(tmpdir)) == set(['layers_folder'])
 
 
-def test_save_layer_multiple_no_named_plugin(
-    tmpdir, layer_data_and_types  # noqa: F811
-):
+def test_save_layer_multiple_no_named_plugin(tmpdir, layer_data_and_types):
     """Test saving multiple layers without naming a plugin."""
     # make writer builtin plugins get called first
     from napari.plugins import plugin_manager
